augmentation/requirements.txt
absl-py==2.1.0
accelerate==0.21.0
aiohappyeyeballs==2.3.7
aiohttp==3.10.4
aiosignal==1.3.1
albumentations==0.5.2
annotated-types==0.7.0
antlr4-python3-runtime==4.8
anyio==4.0.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
attrs==24.2.0
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.3
bleach==6.0.0
blis==0.7.11
boto3==1.28.61
botocore==1.31.61
braceexpand==0.1.7
cachetools==5.5.0
catalogue==2.0.10
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.18.1
cloudpickle==2.2.1
cmake==3.27.0
comm==0.1.4
confection==0.1.5
contourpy==1.1.1
cycler==0.11.0
cymem==2.0.8
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.30.0
distro==1.9.0
dnspython==2.4.1
easydict==1.9
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.1.0
executing==2.0.0
faiss-gpu==1.7.3
fastjsonschema==2.18.0
filelock==3.15.4
fire==0.5.0
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
fvcore==0.1.5.post20221221
gast==0.4.0
google==3.0.0
google-ai-generativelanguage==0.1.0
google-api-core==2.22.0
google-auth==2.34.0
google-auth-oauthlib==1.0.0
google-generativeai==0.1.0rc1
google-pasta==0.2.0
googleapis-common-protos==1.65.0
grpcio==1.65.5
grpcio-status==1.62.3
h11==0.14.0
h5py==3.11.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.24.5
hydra-core==1.1.0
idna==3.7
imageio==2.35.1
imgaug==0.4.0
importlib_metadata==8.2.0
importlib_resources==6.4.3
iniconfig==2.0.0
iopath==0.1.10
IPy==1.1
ipykernel==6.25.2
ipython==8.12.3
ipython-genutils==0.2.0
ipywidgets==8.1.1
jedi==0.19.0
Jinja2==3.1.2
jiter==0.5.0
jmespath==1.0.1
joblib==1.4.2
json5==0.9.14
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.7.0
jupyter-lsp==2.2.0
jupyter_client==8.3.1
jupyter_core==5.3.2
jupyter_server==2.7.3
jupyter_server_terminals==0.4.4
jupyterlab==4.0.6
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.9
jupyterlab_server==2.25.0
keras==2.13.1
kiwisolver==1.4.5
kornia==0.5.0
langcodes==3.4.0
language_data==1.2.0
lazy_loader==0.4
libclang==18.1.1
lit==16.0.6
-e git+https://github.com/facebookresearch/llama.git@6c7fe276574e78057f917549435a2554000a876d#egg=llama
lxml==5.1.0
marisa-trie==1.2.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
nbclient==0.8.0
nbconvert==7.8.0
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.1
notebook==7.0.4
notebook_shim==0.2.3
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu11==8.5.0.96
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu11==10.2.10.91
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu11==11.7.4.91
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu11==2.14.3
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu11==11.7.91
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
omegaconf==2.1.2
openai==1.51.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
opt-einsum==3.3.0
overrides==7.4.0
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
plotly==5.23.0
pluggy==1.0.0
portalocker==2.8.2
preshed==3.0.9
prometheus-client==0.17.1
prompt-toolkit==3.0.39
proto-plus==1.25.0
protobuf==4.25.4
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycocoevalcap==1.2
pycocotools==2.0.7
pycosat==0.6.3
pycparser==2.21
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.16.1
pyparsing==3.0.9
pyre-extensions==0.0.29
pyspark==3.5.0
pytest==7.2.1
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-lightning==1.2.9
pytorch-pretrained-bert==0.6.2
pytz==2024.1
PyWavelets==1.4.1
PyYAML==6.0.2
pyzmq==25.1.1
qtconsole==5.4.4
QtPy==2.4.0
referencing==0.30.2
regex==2024.7.24
requests==2.32.3
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rpds-py==0.10.3
rsa==4.9
ruamel.yaml==0.17.33
ruamel.yaml.clib==0.2.7
s3transfer==0.7.0
safetensors==0.4.4
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.10.1
Send2Trash==1.8.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
shapely==2.0.5
shellingham==1.5.4
six==1.16.0
smart-open==7.0.4
sniffio==1.3.0
soupsieve==2.5
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
sympy==1.12
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.13.1
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.34.0
termcolor==2.3.0
terminado==0.17.1
thinc==8.2.5
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==1.0.8
tinycss2==1.2.1
tokenizers==0.19.1
tomli==2.0.1
torch==2.1.2
torchaudio==2.4.0
torchmetrics==0.2.0
torchvision==0.16.2
tornado==6.3.3
tqdm==4.66.5
traitlets==5.10.1
transformers @ git+https://github.com/huggingface/transformers@61d89c19d853c1e2a597f8dc2bc609988cf2ba4c
triton==2.1.0
typer==0.12.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
ujson==5.9.0
urllib3==2.2.2
warcio==1.7.4
warmup-scheduler==0.3.2
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webdataset==0.2.96
webencodings==0.5.1
Werkzeug==3.0.3
widgetsnbextension==4.0.9
wldhx.yadisk-direct==0.0.6
word2number==1.1
wrapt==1.16.0
xformers==0.0.20
yacs==0.1.8
yarl==1.9.4
zipp==3.20.0
